core: cast Curl proxy port option to long
This fixes this warning with clang:

src/core/core-url.c:1017:5: warning: call to '_curl_easy_setopt_err_long' declared with 'warning' attribute: curl_easy_setopt expects a long argument [-Wattribute-warning]
 1017 |     curl_easy_setopt (curl, CURLOPT_PROXYPORT,
      |     ^
/usr/include/x86_64-linux-gnu/curl/typecheck-gcc.h:50:15: note: expanded from macro 'curl_easy_setopt'
   50 |               _curl_easy_setopt_err_long();                             \
      |               ^

+++ src/core/core-url.c
@@ -1015,7 +1015,7 @@ weeurl_set_proxy (CURL *curl, struct t_proxy *proxy)
 
     /* set proxy port */
     curl_easy_setopt (curl, CURLOPT_PROXYPORT,
-                      CONFIG_INTEGER(proxy->options[PROXY_OPTION_PORT]));
+                      (long)(CONFIG_INTEGER(proxy->options[PROXY_OPTION_PORT])));
 
     /* set username/password */
     if (CONFIG_STRING(proxy->options[PROXY_OPTION_USERNAME])
